+++ app/Http/Controllers/PagesController.php
@@ -35,4 +35,17 @@ class PagesController extends Controller
         $segmen_brosur = DB::table('brosurecatalog')->where('deleted',0)->get();
         return view('pagessales',['brosur' => $segmen_brosur,'profil' => $profil,'lastservice' => $lastservice,'lastservicecount' => $lastservicecount,'cr7data' => $cr7data,'cr7count' => $cr7count]);
     }
+    public function pagesalesconsultation()
+    {
+        $data_user = Auth::user();
+        $profil = DB::table('customerdata')
+                    ->select('*',DB::raw('DATE_FORMAT(customerdata.masa_berlaku_stnk,"%d %M %Y") as berlakustnk'),DB::raw('DATE_FORMAT(customerdata.tanggal_pengerjaan_ssc,"%d %M %Y") as pengerjaanssc'))
+                    ->where('vincode',$data_user->nomor_rangka)->first();
+        $lastservice = DB::table('pkbdata')->where('nomor_rangka',$data_user->nomor_rangka)->orderBy('pkb_date','desc')->first();
+        $lastservicecount = DB::table('pkbdata')->where('nomor_rangka',$data_user->nomor_rangka)->orderBy('pkb_date','desc')->count();
+        $cr7data = DB::table('cr7data')->where('no_polisi',$profil->no_polisi)->orderBy('ID','desc')->first();
+        $cr7count = DB::table('cr7data')->where('no_polisi',$profil->no_polisi)->orderBy('ID','desc')->count();
+        $segmen_brosur = DB::table('brosurecatalog')->where('deleted',0)->get();
+        return view('pagesalesconsultaion',['brosur' => $segmen_brosur,'profil' => $profil,'lastservice' => $lastservice,'lastservicecount' => $lastservicecount,'cr7data' => $cr7data,'cr7count' => $cr7count]);
+    }
 }